Abstract geo admin inline patch from #469 (#489)
