+++ django-rgd-imagery/rgd_imagery/admin/base.py
@@ -1,5 +1,3 @@
-import copy
-
 from django.contrib import admin
 
 # from django.contrib.admin import SimpleListFilter
@@ -9,6 +7,7 @@ from rgd.admin.mixins import (
     SPATIAL_ENTRY_FILTERS,
     TASK_EVENT_FILTERS,
     TASK_EVENT_READONLY,
+    GeoAdminInline,
     _FileGetNameMixin,
     reprocess,
 )
@@ -113,7 +112,7 @@ def convert_images(modeladmin, request, queryset):
 #             return queryset.filter(imagesetspatial__isnull=False)
 
 
-class ImageSetSpatialInline(OSMGeoAdmin, admin.StackedInline):
+class ImageSetSpatialInline(GeoAdminInline):
     model = ImageSetSpatial
     fk_name = 'image_set'
     list_display = (
@@ -122,20 +121,6 @@ class ImageSetSpatialInline(OSMGeoAdmin, admin.StackedInline):
         'created',
     )
     list_filter = MODIFIABLE_FILTERS + SPATIAL_ENTRY_FILTERS
-
-    def __init__(self, parent_model, admin_site):
-        self.admin_site = admin_site
-        self.parent_model = parent_model
-        self.opts = self.model._meta
-        self.has_registered_model = admin_site.is_registered(self.model)
-        overrides = copy.deepcopy(admin.options.FORMFIELD_FOR_DBFIELD_DEFAULTS)
-        for k, v in self.formfield_overrides.items():
-            overrides.setdefault(k, {}).update(v)
-        self.formfield_overrides = overrides
-        if self.verbose_name is None:
-            self.verbose_name = self.model._meta.verbose_name
-        if self.verbose_name_plural is None:
-            self.verbose_name_plural = self.model._meta.verbose_name_plural
 
 
 @admin.register(ImageSet)
